+++ xposed-module/app/src/main/java/com/specter/module/gen/IdentityService.java
@@ -1,4 +1,4 @@
-package com.fleet.idrotate.gen;
+package com.specter.module.gen;
 
 import android.content.Context;